+++ be/src/main/java/com/foodymoody/be/comment/domain/CommentValidator.java
@@ -1,4 +1,4 @@
-package com.foodymoody.be.comment.util;
+package com.foodymoody.be.comment.domain;
 
 import static com.foodymoody.be.common.util.Constants.UTILITY_CLASS;
